Added LogoKill to project Fixed bugs in OSD Added black level and white level in gamma for Dan to test Made new Wait for vertical blank settting to try and spped up OSD painting Added check for Custom channel list
git-svn-id: https://svn.code.sf.net/p/deinterlace/code/trunk/DScaler@715 6e12ab15-cf2f-4ffe-bf36-45892a6a5b89

+++ DScaler/OutThreads.cpp
@@ -674,8 +674,8 @@ DWORD WINAPI YUVOutThread(LPVOID lpThreadParameter)
 			{
 				nTotalDropFields += Timing_GetDroppedFields();
 				nTotalUsedFields += Timing_GetUsedFields();
-				nDropFieldsLastSec = Timing_GetDroppedFields() * 1000 / (double)(CurrentTickCount - dwLastSecondTicks);
-				nUsedFieldsLastSec = Timing_GetUsedFields() * 1000 / (double)(CurrentTickCount - dwLastSecondTicks);
+				nDropFieldsLastSec = (double)Timing_GetDroppedFields() * 1000.0 / (double)(CurrentTickCount - dwLastSecondTicks);
+				nUsedFieldsLastSec = (double)Timing_GetUsedFields() * 1000.0 / (double)(CurrentTickCount - dwLastSecondTicks);
 				Timing_ResetDroppedFields();
 				Timing_ResetUsedFields();
 				nSecTicks += CurrentTickCount - dwLastSecondTicks;
